+++ src/database/initialization/organization_to_json.sql
@@ -3,7 +3,8 @@ RETURNS JSONB AS $$
 BEGIN
   RETURN jsonb_build_object(
     'id', organization.id,
-    'employerIdentificationNumber', organization.employer_identification_number,
+    'taxId', organization.tax_id,
+    'employerIdentificationNumber', organization.tax_id,
     'name', organization.name,
     'createdAt', organization.created_at
   );